#13 Sistema de grupos

+++ components/groups/GroupsList.jsx
@@ -7,11 +7,13 @@ import {Field, Form, Formik} from "formik";
 import * as Yup from "yup";
 import {useRouter} from "next/router";
 import DeleteIcon from '@mui/icons-material/Delete';
+import sweetAlert from 'sweetalert'
 
 const GroupsList = ({userId}) => {
 
     const [groups, setGroups] = useState([]);
     const [dialogOpen, setDialogOpen] = useState(false);
+    const [initialValues, setInitialValues] = useState(CreateGroupSchema.default())
     const router = useRouter();
 
     const fetchUserGroups = async () => {
@@ -50,13 +52,21 @@ const GroupsList = ({userId}) => {
         } catch (e) {
             console.error(e);
         }
+    }
 
+    const editGroup = async (group) => {
+        console.log('grugru', group)
+        setInitialValues(group);
+        setDialogOpen(true);
     }
 
     return (
         <div className={`${styles.GroupListBody} ${buttonStyle.cardWrapper}`}>
             <div className={styles.CreateGroupWrapper}>
-                <Button variant="outlined" onClick={() => setDialogOpen(true)} className={styles.submitButton}>
+                <Button variant="outlined" onClick={() => {
+                    setInitialValues(CreateGroupSchema.default());
+                    setDialogOpen(true)
+                }} className={styles.submitButton}>
                     Criar grupo
                 </Button>
                 {groups.length === 0 &&
@@ -65,15 +75,16 @@ const GroupsList = ({userId}) => {
                 {groups.length !== 0 &&
                     groups.map(group => {
                         return (
-                            <GroupItem group={group} fetchUserGroups={fetchUserGroups} />
+                            <GroupItem group={group} fetchUserGroups={fetchUserGroups} editGroup={editGroup} />
                             )
                     })
                 }
             </div>
             <Dialog open={dialogOpen} onClose={() => setDialogOpen(false)}>
                 <DialogTitle>Criar grupo</DialogTitle>
-                <Formik initialValues={CreateGroupSchema.default()} onSubmit={createGroup}>
+                <Formik initialValues={initialValues} onSubmit={createGroup}>
                     {({values, handleChange, handleBlur, errors, touched}) => {
+                        console.log('vavavava', values, initialValues)
                         return (
                             <Form className={`${styles.groupForm} ${buttonStyle.cardWrapper}`}>
                                 <Field
@@ -102,16 +113,30 @@ const CreateGroupSchema = Yup.object().shape({
     name: Yup.string().required('É necessário fornecer um nome').default(''),
 });
 
-const GroupItem = ({group, fetchUserGroups}) => {
+const GroupItem = ({group, fetchUserGroups, editGroup}) => {
 
     const deleteGroup = async () => {
         try {
-            await fetch('/api/userGroups', {
-                method: 'DELETE',
-                headers: { 'Content-Type': 'application/json' },
-                body: JSON.stringify({groupId: group.id}),
-            });
-            await fetchUserGroups();
+
+            sweetAlert({
+                title: "Atenção",
+                text: "Quer mesmo deletar este grupo?",
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+            })
+                .then(async (willDelete) => {
+                    if (willDelete) {
+                        await fetch('/api/userGroups', {
+                            method: 'DELETE',
+                            headers: { 'Content-Type': 'application/json' },
+                            body: JSON.stringify({groupId: group.id}),
+                        });
+                        await fetchUserGroups();
+                    }
+                });
+            
+
         } catch (e) {
             console.error(e);
         }
@@ -124,7 +149,7 @@ const GroupItem = ({group, fetchUserGroups}) => {
               <span>Membros: {group.members.length}</span>
               <div>
                   <IconButton><Icon>east</Icon></IconButton>
-                  <IconButton><Icon>edit</Icon></IconButton>
+                  <IconButton onClick={() => editGroup(group)}><Icon>edit</Icon></IconButton>
                   <IconButton onClick={deleteGroup}><DeleteIcon /></IconButton>
               </div>
           </div>
